move encrypted_file and test_helpers into a spec/shared folder
- update requirements in other files
