+++ .env
@@ -1,4 +1,4 @@
-ENCRYPTED_FILE_PATH='./spec/data/encrypted_file.gpg'
+ENCRYPTED_FILE_PATH='./spec/shared/encrypted_file.gpg'
 DECRYPTED_FILE_PATH='./spec/data/generated_decrypted_file.txt'
 PGP_PRIVATE_KEY_PATH='./spec/data/generated_secret_key_file.gpg'
 PRIVATE_KEY="-----BEGIN PGP PRIVATE KEY BLOCK-----\n\nlQIGBFrADhMBBADl+8n6lZD7MWVCzAEoPv4CQw1oz3CXt44KDvNiA3X7HtGNGaue\n/Z2wJtCF/AU7q9y42Y3SDkS+JywtSaUgFAU13eSX2VXyP3JsPewfbnsd7AFZZYIp\n0ZVQGOdezCHDngJ4bn6GYEF1QWUTwrd8Z3PBegq/o7w7WlOyedBAFXrrowARAQAB\n/gcDAptP+zuPpSK4YMpju+PLpIAfN8eU2RxGDr2fuGf/DpeebeLN7FB4c4mwsOv1\nVg5CDjggMwan7TAvjOsXHvvo5wGhlWfNvwotBl1IW8Vrfs0a2q+KDtzQ/k/NWR4c\n0CnUlU3XSMMKj5Knz8EeqCVVJRgjR2ndTDHo0OSFI7V0Sd7SSrakxFi3S5SHQ2oi\nQsKTMWlgAKM8J7twWtcYGLZ0afsOxksMDsb5NLBUmb2oK/RlqQboLIklFg730rUx\nKTsPjIAYcCkJpiYBvW/mb9tW5o2zU8JLXo1426R59XpQzbv0MaiJTwIvRlIocHBk\nlbim7IDQosFxcS6ILI27liXLFLzXiSfQRoov0u3foD5yLPNCOKlZTr9pTiuC9z8p\nhcObTRJ3Fuj+aUw4UYwvi2YLIEDw5yzmJAi7dfwkScCBVjJ/Vy5zsGmuOlf8QPQu\nb/aaqWwJw7MkoBunNoQWa9K3/7VUpw2NuRkjOH2Nmnq1immpzM7MW/K0KEdQR2Vu\naWUgKGZvciB0ZXN0KSA8Z3BnZW5pZUBleGFtcGxlLmNvbT6IuAQTAQIAIgUCWsAO\nEwIbAwYLCQgHAwIGFQgCCQoLBBYCAwECHgECF4AACgkQYiTYJ3FfIC+bjwQAgMmy\nOSAY5Qmy7mtniRrgGiZ7d3mYN7E8+BCJj4GemmE2J1sGdujE3z6tIiGWRCwJvdym\nzgLenI3RdvlRM+mdFOjJvExsft0fRopRpS4mf/NojcItHaVusm3CO3VOT55r81A5\nphZ9/AkQUbQw7N5iciJWJvF+QZrOwuU9/NdHM0mdAgUEWsAOEwEEALd7mirIcYcT\nAnlSCfNkRC4A52FWy1ON4DxftzQikJnmNe2VeHaIo/lFeJGFDUQ0LXtAJGjtOcOI\n3kZNzfC7zLzyG6guRS2Jy9OOaDtegbcFfdxoCVGKT0hPokmis8h5fpvTu/6utA5Q\nrkuOxwG68Gf5QIYQwqCoRqKZh4QQvPhdABEBAAH+BwMCm0/7O4+lIrhgwcJho2iA\nItNeYUEXnHLGWZerDg8dPB5MBcgQ1F1lILk53PNZK/LHp9arkwgvIejvyOAiqyty\nRNH03ibsOXoQhiojnTQhz/eHw9wyG7zs2vR95xatXm0jkUhesxOmqQ2bO+k1lsIC\nnp1zNWRL+tG8QQQn0Q1OffknNAedatq0CvPrb3V50npTyNpSBLXpKWH7TrIYWWL0\nPVgWT3nyvSh67w2N6d+/j/5cyYYU7cz214awTMF5Wd4emjYjQnke58G1bPW76EKx\nhyGv2PUctIxx4YudC1MojFwqMxdgl+1HFUaP0KBXu+Rr61h2qSPMRPD91LfaFQ02\nWZeI4bb0CwmdP7SkyfCrKTssAteLRBxhvF4iDvMJcNTbvZWpZks0Ce//+IeX69BI\n4nJFTSD+nalajJ+6qAk5pR+1oRiCcRtD2bd0R6sQ5WEu82PlSCuljsJ1H+D8sCP5\nGSl0sNTOXCiHtc9v7g0twetCZTfDiJ8EGAECAAkFAlrADhMCGwwACgkQYiTYJ3Ff\nIC/QFgP/dx7z7Du55iUO7dJPmF31H6BUoR/thwiZohnjq+Oj+y+bkjTVMqSijvEe\n7YIX715V0Sf0mkhpo3XELEE7Plf9eliVV4s8dEN8OsblG0vVB3wPYagLVBrBIAF6\nVmxH2nmF05s6B6giZxa6yokIyZYJG38+uUUp3fpX3JIdWzki9YU=\n=loGB\n-----END PGP PRIVATE KEY BLOCK-----"
