+++ spec/process_spec.rb
@@ -4,7 +4,7 @@ require 'dotenv/load'
 require 'rspec'
 require 'fileutils'
 require './spec/data/variables'
-require './spec/test_helpers.rb'
+require './spec/shared/test_helpers.rb'
 require './lib/gnu_pg/process.rb'
 require './lib/gnu_pg/import_key.rb'
 require './lib/gnu_pg/decrypt.rb'
