+++ spec/shared/encrypted_file.gpg
