+++ spec/shared/test_helpers.rb
